IAT-440: Refactor header and footer components into shared module.
