+++ src/app/app.module.ts
@@ -10,26 +10,21 @@ import {Ng2BootstrapModule} from "ngx-bootstrap/ng2-bootstrap";
 
 import {AppRoutingModule} from "./app.routing";
 import {AppComponent} from "./app.component";
-import {HeaderComponent} from "./header/header.component";
-import {FooterComponent} from "./footer/footer.component";
 import {ItemSelectTypeComponent} from "./item/item-create/item-select-type/item-select-type.component";
-import {HomeComponent} from "./home/home.component";
 import {ItemCreateComponent} from "./item/item-create/item-create.component";
 import {ItemLoadSaComponent} from "./item/item-load-sa/item-load-sa.component";
 import {ItemLoadComponent} from "./item/item-load/item-load.component";
 import {ItemRedirectComponent} from "./item/item-redirect/item-redirect.component";
 import {NoRouteComponent} from "./no-route/no-route.component";
-import {Logger} from "./core/logger.service";
 import {ToastyModule} from "ng2-toasty";
-import {AlertService} from "./core/alert.service";
-import {ItemSearchComponent} from './item/item-search/item-search.component';
+import {ItemSearchComponent} from "./item/item-search/item-search.component";
 import {CoreModule} from "./core/core.module";
+import {SharedModule} from "./shared/shared.module";
+import {HomeComponent} from "./home/home.component";
 
 @NgModule({
   declarations: [
     AppComponent,
-    FooterComponent,
-    HeaderComponent,
     HomeComponent,
     ItemCreateComponent,
     ItemLoadSaComponent,
@@ -50,6 +45,7 @@ import {CoreModule} from "./core/core.module";
     ModalModule.forRoot(),
     Ng2BootstrapModule.forRoot(),
     ReactiveFormsModule,
+    SharedModule,
     ToastyModule.forRoot(),
   ],
   bootstrap: [
