--- src/app/footer/footer.component.html
@@ -1,7 +1,0 @@
-<footer>
-  <div class="container">
-    <p class="pull-right">
-      SmarterApp - Item Authoring Tool <span *ngIf="buildInfo">v{{buildInfo?.build.version}}</span>
-    </p>
-  </div>
-</footer>
