+++ src/app/shared/footer/footer.component.html
@@ -1,0 +1,7 @@
+<footer>
+  <div class="container">
+    <p class="pull-right">
+      SmarterApp - Item Authoring Tool <span *ngIf="buildInfo">v{{buildInfo?.build?.version}}</span>
+    </p>
+  </div>
+</footer>
