+++ src/app/shared/footer/footer.component.less
