+++ src/app/shared/header/header.component.html
@@ -1,15 +1,22 @@
 <header>
   <div class="container">
-
     <div class="row">
       <div class="col-lg-12">
         <nav class="navbar navbar-default" role="navigation">
+
+          <!-- App title -->
           <div class="navbar-header">
-            <a class="navbar-brand" routerLink="/" title="Reporting System Dashboard"><img
-              src="/assets/images/SmarterBalanced-Logo.png"
-              alt="Smarter Balanced Assessment Consortium"></a>
+            <!-- App logo -->
+            <a class="navbar-brand" routerLink="/" title="Reporting System Dashboard">
+              <img src="/assets/images/SmarterBalanced-Logo.png"
+                   alt="Smarter Balanced Assessment Consortium">
+            </a>
+
+            <!-- App name -->
             <span class="application-title">Item Authoring Tool</span>
           </div>
+
+          <!-- User name drop-down -->
           <div>
             <ul class="nav navbar-nav navbar-right">
               <li class="dropdown" dropdown>
@@ -31,21 +38,26 @@
               </li>
             </ul>
           </div>
+
         </nav>
       </div>
     </div>
   </div>
 </header>
 
-<!-- Modal -->
-<div id="myModal" class="modal fade" bsModal #staticModal="bs-modal" [config]="{backdrop: 'static'}"
-     tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
-
+<!-- Log out confirmation dialog -->
+<div id="myModal"
+     class="modal fade"
+     bsModal
+     #staticModal="bs-modal"
+     [config]="{backdrop: 'static'}"
+     tabindex="-1"
+     role="dialog"
+     aria-labelledby="myModalLabel">
   <div class="modal-dialog" role="document">
     <div class="modal-content">
       <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close"
-                (click)="staticModal.hide()">
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close" (click)="staticModal.hide()">
           <span aria-hidden="true">&times;</span>
         </button>
         <h3 class="modal-title" id="myModalLabel">Log Out?</h3>
