+++ src/app/shared/shared.module.ts
@@ -1,0 +1,22 @@
+import {NgModule} from "@angular/core";
+import {CommonModule} from "@angular/common";
+import {FooterComponent} from "./footer/footer.component";
+import {HeaderComponent} from "./header/header.component";
+import {Ng2BootstrapModule} from "ngx-bootstrap";
+
+@NgModule({
+  imports: [
+    CommonModule,
+    Ng2BootstrapModule
+  ],
+  declarations: [
+    FooterComponent,
+    HeaderComponent
+  ],
+  exports: [
+    FooterComponent,
+    HeaderComponent
+  ]
+})
+export class SharedModule {
+}
